# APP_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: VPN Panel Android App API
  version: 1.0.0
  description: Invitation-only Android VPN client API. Native App uses Bearer tokens only.
servers:
  - url: https://panel.example.com/api/v1
paths:
  /:
    get:
      summary: Public API configuration
      responses:
        '200':
          description: API status and registration configuration
  /health:
    get:
      summary: Health check
      responses:
        '200': {description: Healthy}
  /register:
    post:
      summary: Register with invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code, username, password]
              properties:
                invite_code: {type: string}
                username: {type: string, minLength: 3, maxLength: 32}
                password: {type: string, minLength: 8}
      responses:
        '201': {description: Registered}
        '403': {description: Registration closed or invalid invite}
        '409': {description: Username already exists}
        '429': {description: Rate limited}
  /login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        '200':
          description: Login token
        '401': {description: Invalid credentials}
        '403': {description: Account disabled}
        '429': {description: Rate limited}
  /app/bootstrap:
    get:
      summary: App cold-start bootstrap data
      security: [{bearerAuth: []}]
      responses:
        '200': {description: User and all enabled nodes grouped by country}
        '401': {description: Token invalid or expired}
        '403': {description: Account disabled}
  /me:
    get:
      summary: Current user state
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current active user}
  /nodes:
    get:
      summary: All enabled nodes grouped and sorted by country
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Node payload}
  /change-password:
    post:
      summary: Change user password and rotate all tokens
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: {type: string}
                new_password: {type: string, minLength: 8}
      responses:
        '200': {description: Password changed and new token issued}
        '401': {description: Current password invalid}
  /logout:
    post:
      summary: Revoke current token
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Logged out}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque